% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{comparing_pm10}
\alias{comparing_pm10}
\title{shinyapp "pm10 comparing"}
\usage{
comparing_pm10()
}
\value{
Runs the shiny app containing an interactive analysis on the data contained in the input database.
}
\description{
The function does not require an input variable. It depends on the input dataset
}
